pyproject.toml
[project]
name = "sky-edge"
version = "0.1.0"
description = "sky-edge is an opinionated API client for Blackbaud Raiser's Edge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]

[project.scripts]
sky-edge = "sky_edge:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
